[pax] Updated to patch 4.2.1-test2
Updated to pax-linux-4.2.1-test2.patch

PaX Team says:
  - Emese fixed a false positive integer truncation detected in ide_probe_port, reported by spender
  - fixed a few unnecessary integer truncations detected by the new size overflow plugin, there're still some outstanding issues though:
    - http://marc.info/?l=linux-usb&m=144327802931697&w=2
    - http://lists.infradead.org/pipermail/linux-mtd/2015-September/062130.html
    - http://news.gmane.org/gmane.linux.kernel.cifs (will show up later hopefully)

+++ fs/binfmt_elf.c
@@ -1836,7 +1836,7 @@ static void fill_elf_header(struct elfhdr *elf, int segs,
 	return;
 }
 
-static void fill_elf_note_phdr(struct elf_phdr *phdr, int sz, loff_t offset)
+static void fill_elf_note_phdr(struct elf_phdr *phdr, size_t sz, loff_t offset)
 {
 	phdr->p_type = PT_NOTE;
 	phdr->p_offset = offset;
